# src/docs/swagger.yaml
basePath: /api/v1
definitions:
  model.Response:
    properties:
      attributionHTML:
        type: string
      attributionText:
        type: string
      code:
        type: integer
      copyright:
        type: string
      data:
        type: string
      etag:
        type: string
      status:
        type: string
    type: object
  model.ResponseError:
    properties:
      code:
        type: integer
      status:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server marvel server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Marvel Example API
  version: "1.0"
paths:
  /character:
    get:
      consumes:
      - application/json
      description: get characters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseError'
            type: object
      summary: List characters
      tags:
      - characters
  /characters/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Character ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseError'
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseError'
            type: object
      summary: Show a character
      tags:
      - characters
swagger: "2.0"